requirements.txt
accelerate>=0.24.1
auto-gptq==0.3.0
deepspeed==0.8.3
datasets>=2.10.1
flash-attn
jsonlines
peft>=0.5.0
safetensors>=0.3.1
torch==1.13.1
transformers==4.30.0
fastapi>=0.109.1
uvicorn==0.17.6
streamlit>=1.30.0
pyyaml>=6.0
psutil>=5.9.0
requests>=2.28.0
